Address review: unify filter logic, remove dead variable, smooth zoom pill
- Move labelMatchesFilter() + ANIMAL_LABELS to TrackingState.kt as
  single source of truth (was duplicated in ViewModel + CameraScreen)
- Remove dead rawDisplayObjects intermediate variable
- Zoom pill always composed, uses alpha for smooth fade instead of
  conditional show/hide
