+++ app/src/main/java/com/haptictrack/tracking/TrackingState.kt
@@ -22,6 +22,18 @@ enum class TrackingFilter {
     NON_PERSON_ONLY
 }
 
+private val ANIMAL_LABELS = setOf(
+    "cat", "dog", "bird", "horse", "sheep", "cow",
+    "elephant", "bear", "zebra", "giraffe"
+)
+
+fun labelMatchesFilter(label: String?, filter: TrackingFilter): Boolean = when (filter) {
+    TrackingFilter.ALL -> true
+    TrackingFilter.PERSON_ONLY -> label == "person"
+    TrackingFilter.PETS -> label in ANIMAL_LABELS
+    TrackingFilter.NON_PERSON_ONLY -> label != "person"
+}
+
 data class TrackedObject(
     val id: Int,
     val boundingBox: RectF,
